# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sanjeevani Backend",
    "description": "Sanjeevani supply-chain disruption system backend.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/disruptions": {
      "get": {
        "tags": [
          "disruptions"
        ],
        "summary": "List Disruptions",
        "operationId": "list_disruptions_api_v1_disruptions_get",
        "parameters": [
          {
            "name": "stage",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/DisruptionStage"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stage"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisruptionList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/disruptions/{disruption_id}": {
      "get": {
        "tags": [
          "disruptions"
        ],
        "summary": "Get Disruption",
        "operationId": "get_disruption_api_v1_disruptions__disruption_id__get",
        "parameters": [
          {
            "name": "disruption_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Disruption Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Disruption"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/disruptions/{disruption_id}/impact": {
      "get": {
        "tags": [
          "disruptions"
        ],
        "summary": "Get Disruption Impact",
        "operationId": "get_disruption_impact_api_v1_disruptions__disruption_id__impact_get",
        "parameters": [
          {
            "name": "disruption_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Disruption Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImpactGraph"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/disruptions/{disruption_id}/plan": {
      "get": {
        "tags": [
          "disruptions"
        ],
        "summary": "Get Disruption Plan",
        "description": "Latest remediation plan, shaped as the two-sided diff the frontend's\nPlanDiffPanel renders: incumbent PO lines on the left, the solver's\nproposed allocation on the right.",
        "operationId": "get_disruption_plan_api_v1_disruptions__disruption_id__plan_get",
        "parameters": [
          {
            "name": "disruption_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Disruption Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanDiffResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/disruptions/simulate": {
      "post": {
        "tags": [
          "simulate"
        ],
        "summary": "Simulate Disruption",
        "operationId": "simulate_disruption_api_v1_disruptions_simulate_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/demo/reset": {
      "post": {
        "tags": [
          "demo"
        ],
        "summary": "Demo Reset",
        "operationId": "demo_reset_api_v1_demo_reset_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoResetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/demo/state": {
      "get": {
        "tags": [
          "demo"
        ],
        "summary": "Demo State",
        "operationId": "demo_state_api_v1_demo_state_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DemoResetResponse": {
        "properties": {
          "cleared": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Cleared"
          },
          "reseeded_disruptions": {
            "type": "integer",
            "title": "Reseeded Disruptions"
          },
          "elapsed_ms": {
            "type": "number",
            "title": "Elapsed Ms"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "cleared",
          "reseeded_disruptions",
          "elapsed_ms"
        ],
        "title": "DemoResetResponse"
      },
...
      "DemoState": {
        "properties": {
          "disruption_count_by_stage": {
            "items": {
              "$ref": "#/components/schemas/DemoStageCount"
            },
            "type": "array",
            "title": "Disruption Count By Stage"
          },
          "integrations": {
            "additionalProperties": {
              "$ref": "#/components/schemas/IntegrationStatus"
            },
            "type": "object",
            "title": "Integrations"
          },
          "ttm_loaded": {
            "type": "boolean",
            "title": "Ttm Loaded"
          },
          "ws_client_count": {
            "type": "integer",
            "title": "Ws Client Count"
          },
          "db_roundtrip_ms": {
            "type": "number",
            "title": "Db Roundtrip Ms"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "disruption_count_by_stage",
          "integrations",
          "ttm_loaded",
          "ws_client_count",
          "db_roundtrip_ms",
          "updated_at"
        ],
        "title": "DemoState"
      },
...
      "Disruption": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "$ref": "#/components/schemas/DisruptionType"
          },
          "stage": {
            "$ref": "#/components/schemas/DisruptionStage"
          },
          "detected_at": {
            "type": "string",
            "title": "Detected At"
          },
          "vendor": {
            "$ref": "#/components/schemas/VendorRef"
          },
          "affected_po_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Affected Po Ids"
          },
          "headline": {
            "type": "string",
            "title": "Headline"
          },
          "exposure": {
            "$ref": "#/components/schemas/Exposure"
          },
          "diagnosis": {
            "$ref": "#/components/schemas/Diagnosis"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/SourcingCandidate"
            },
            "type": "array",
            "title": "Candidates"
          },
          "approval": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Approval"
              },
              {
                "type": "null"
              }
            ]
          },
          "negotiation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Negotiation"
              },
              {
                "type": "null"
              }
            ]
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/TimelineEvent"
            },
            "type": "array",
            "title": "Timeline"
          },
          "detector_source": {
            "$ref": "#/components/schemas/DetectorSource"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "type",
          "stage",
          "detected_at",
          "vendor",
          "affected_po_ids",
          "headline",
          "exposure",
          "diagnosis",
          "candidates",
          "approval",
          "negotiation",
          "timeline",
          "detector_source"
        ],
        "title": "Disruption"
      },
      "DisruptionList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DisruptionSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "DisruptionList"
      },
      "DisruptionStage": {
        "type": "string",
        "enum": [
          "DETECTED",
          "DIAGNOSED",
          "SOURCING",
          "AWAITING_APPROVAL",
          "APPROVED",
          "REJECTED",
          "NEGOTIATING",
          "NEGOTIATED",
          "SETTLEMENT_PENDING",
          "SETTLED",
          "CLOSED",
          "FAILED"
        ],
        "title": "DisruptionStage"
      },
      "DisruptionSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "$ref": "#/components/schemas/DisruptionType"
          },
          "stage": {
            "$ref": "#/components/schemas/DisruptionStage"
          },
          "detected_at": {
            "type": "string",
            "title": "Detected At"
          },
          "vendor": {
            "$ref": "#/components/schemas/VendorRef"
          },
          "headline": {
            "type": "string",
            "title": "Headline"
          },
          "exposure_total_paise": {
            "type": "integer",
            "title": "Exposure Total Paise"
          },
          "exposure_total_display": {
            "type": "string",
            "title": "Exposure Total Display"
          },
          "detector_source": {
            "$ref": "#/components/schemas/DetectorSource"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "type",
          "stage",
          "detected_at",
          "vendor",
          "headline",
          "exposure_total_paise",
          "exposure_total_display",
          "detector_source"
        ],
        "title": "DisruptionSummary",
        "description": "Lighter shape used in list views."
      },
      "DisruptionType": {
        "type": "string",
        "enum": [
          "DELIVERY_DELAY",
          "VENDOR_UNRESPONSIVE",
          "QUALITY_REJECTION",
          "STOCKOUT_RISK",
          "PRICE_SHOCK",
          "LOGISTICS_HOLD"
        ],
        "title": "DisruptionType"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImpactGraph": {
        "properties": {
          "disruption_id": {
            "type": "string",
            "title": "Disruption Id"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/ImpactNode"
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/ImpactEdge"
            },
            "type": "array",
            "title": "Edges"
          },
          "summary": {
            "$ref": "#/components/schemas/ImpactSummary"
          },
          "computed_at": {
            "type": "string",
            "title": "Computed At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "disruption_id",
          "nodes",
          "edges",
          "summary",
          "computed_at"
        ],
        "title": "ImpactGraph"
      },
...
      "PlanDiffResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "disruption_id": {
            "type": "string",
            "title": "Disruption Id"
          },
          "changes": {
            "items": {
              "$ref": "#/components/schemas/PlanChangeDetail"
            },
            "type": "array",
            "title": "Changes"
          },
          "exposure_before_paise": {
            "type": "integer",
            "title": "Exposure Before Paise"
          },
          "exposure_before_display": {
            "type": "string",
            "title": "Exposure Before Display"
          },
          "exposure_after_paise": {
            "type": "integer",
            "title": "Exposure After Paise"
          },
          "exposure_after_display": {
            "type": "string",
            "title": "Exposure After Display"
          },
          "cost_to_resolve_paise": {
            "type": "integer",
            "title": "Cost To Resolve Paise"
          },
          "cost_to_resolve_display": {
            "type": "string",
            "title": "Cost To Resolve Display"
          },
          "net_saving_paise": {
            "type": "integer",
            "title": "Net Saving Paise"
          },
          "net_saving_display": {
            "type": "string",
            "title": "Net Saving Display"
          },
          "requires_escalation": {
            "type": "boolean",
            "title": "Requires Escalation"
          },
          "escalation_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Escalation Reason"
          },
          "solver": {
            "type": "string",
            "title": "Solver"
          },
          "solve_ms": {
            "type": "number",
            "title": "Solve Ms"
          }
        },
        "type": "object",
        "required": [
          "id",
          "disruption_id",
          "changes",
          "exposure_before_paise",
          "exposure_before_display",
          "exposure_after_paise",
          "exposure_after_display",
          "cost_to_resolve_paise",
          "cost_to_resolve_display",
          "net_saving_paise",
          "net_saving_display",
          "requires_escalation",
          "solver",
          "solve_ms"
        ],
        "title": "PlanDiffResponse"
      },
...
      "SimulateRequest": {
        "properties": {
          "scenario": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario"
          },
          "vendor_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendor Id"
          },
          "kind": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScenarioKind"
              },
              {
                "type": "null"
              }
            ]
          },
          "effective_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Date"
          }
        },
        "type": "object",
        "title": "SimulateRequest"
      },
      "SimulateResponse": {
        "properties": {
          "disruption_id": {
            "type": "string",
            "title": "Disruption Id"
          },
          "scenario": {
            "type": "string",
            "title": "Scenario"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "newly_triggered": {
            "type": "boolean",
            "title": "Newly Triggered"
          }
        },
        "type": "object",
        "required": [
          "disruption_id",
          "scenario",
          "stage",
          "newly_triggered"
        ],
        "title": "SimulateResponse"
      },
...
    }
  }
}